Experimental cache approach

+++ NetworkingPlayground/ObjectSerialization/Source/Code/ObjectSerialization/Streams/OutputMemoryBitStream.cpp
@@ -76,7 +76,7 @@ namespace Serialization { namespace Stream {
 	void OutputMemoryBitStream::WriteFreeBits(uint8_t inData, size_t inBitCount)
 	{
 		const uint32_t byteOffset = mBitHead >> 3;
-		const uint32_t bitOffset = mBitHead & 0x7;
+		const uint8_t bitOffset = mBitHead & 0x7;
 
 		assert(bitOffset + inBitCount <= 8);
